Initial pass of a step funciton for the core
Create an offset type to more easily add and subtract offsets within the
context of a core, plus some helpers for pulling values out of fields and
initial logic scaffolding for step() itself

+++ corewars-core/src/load_file/program.rs
@@ -6,7 +6,7 @@ use std::{collections::HashMap, fmt};
 use super::{Instruction, PseudoOpcode, UOffset};
 
 pub type Instructions = Vec<Instruction>;
-pub type LabelMap = HashMap<String, usize>;
+pub type LabelMap = HashMap<String, u32>;
 
 /// A parsed Redcode program, which can be loaded into a core for execution
 #[derive(Default, PartialEq)]
